+++ gamja/src/components/common/CategoryBar.jsx
