+++ gamja/src/components/common/ChatCard.jsx
