+++ gamja/src/components/common/MypageCard.jsx
