+++ gamja/src/components/common/QuoteSection.jsx
