+++ gamja/src/components/common/SearchCard.jsx
