+++ gamja/src/pages/mentor/ConcernsPage.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import TopBar from "../components/TopBar";
-import CategoryBar from "../components/CategoryBar";
-import Concern from "../components/Concern";
+import TopBar from "../../components/common/TopBar";
+import CategoryBar from "../../components/common/CategoryBar";
+import Concern from "../../components/mentor/Concern";
 
 const initConcernList = [
   {
